backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pdfplumber
python-docx
google-generativeai
